update-20220722

+++ src/main/java/com/github/capture/model/TcpPacketRecord.java
@@ -48,7 +48,7 @@ public class TcpPacketRecord implements Serializable {
 
     private short ipFragmentOffset;
 
-    private byte ipTtl;
+    private int ipTtl;
 
     private byte ipProtocol;
